+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Upgrade mysql to Version 8.0.32 (#10)
+- Do not use flag `--log-warnings` anymore as it has been removed in mysql8
+  - Use `log_error_verbosity` instead and set default log level to WARN.
 
 ## [v5.7.37-4] - 2022-06-16
 ### Changed